mcp_cli/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-cli"
version = "1.0.0"
description = "MCP (Model Context Protocol) CLI client"
authors = [
    {name = "Grok", email = "grok@x.ai"},
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.0.0",
    "requests>=2.25.0",
    "sseclient-py>=1.7.0",
    "pydantic>=2.0.0",
    "rich>=10.0.0",
    "pyyaml>=6.0.0",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
]

[project.scripts]
mcp-cli = "mcp_cli.cli:cli"

[tool.setuptools]
packages = ["mcp_cli"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"